break geometry_msgs dep
