+++ tf2/CMakeLists.txt
@@ -16,7 +16,6 @@ option(TF2_ROS_BUILD "Build TF2 as a ROS package" ON)
 if(TF2_ROS_BUILD)
   find_package(ament_cmake_ros REQUIRED)
   find_package(builtin_interfaces REQUIRED)
-  find_package(geometry_msgs REQUIRED)
   find_package(rcutils REQUIRED)
   find_package(rosidl_runtime_cpp REQUIRED)
 endif()
