+++ tf2_eigen/package.xml
@@ -18,11 +18,12 @@
 
   <depend>geometry_msgs</depend>
   <depend>tf2</depend>
-  <depend>tf2_ros</depend>
+  <depend>tf2_geometry_msgs</depend>
 
   <test_depend>ament_cmake_gtest</test_depend>
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
+  <test_depend>tf2_ros</test_depend>
 
   <build_export_depend>eigen</build_export_depend>
