+++ tf2_eigen_kdl/package.xml
@@ -24,10 +24,11 @@
 
   <depend>orocos_kdl_vendor</depend>
   <depend>tf2</depend>
-
+  
   <test_depend>ament_cmake_gtest</test_depend>
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
+  <test_depend>tf2_geometry_msgs</test_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
