+++ tf2_eigen_kdl/test/tf2_eigen_kdl_test.cpp
@@ -34,6 +34,8 @@
 #include <tf2/convert.h>
 #include <gtest/gtest.h>
 
+#include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
+
 using Vector6d = Eigen::Matrix<double, 6, 1>;
 
 TEST(TfEigenKdl, TestRotationQuaternion)
