+++ tf2_ros/package.xml
@@ -26,6 +26,7 @@
   <depend>rclcpp_components</depend>
   <depend>tf2</depend>
   <depend>tf2_msgs</depend>
+  <depend>tf2_geometry_msgs</depend>
 
   <test_depend>ament_cmake_gtest</test_depend>
   <test_depend>ament_lint_auto</test_depend>
