+++ tf2_ros/src/buffer.cpp
@@ -40,6 +40,8 @@
 #include <string>
 #include <thread>
 
+#include "tf2_geometry_msgs/tf2_geometry_msgs.hpp"
+
 namespace tf2_ros
 {
 inline
@@ -65,7 +67,7 @@ BufferCoreROSConversions::lookupTransform(
     target_frame,
     source_frame, time);
 
-  geometry_msgs::msg::TransformStamped msg = tf2::toMsg<tf2::Stamped<tf2::Transform>, geometry_msgs::msg::TransformStamped>(stamped_transform);
+  geometry_msgs::msg::TransformStamped msg = tf2::toMsg(stamped_transform);
   msg.child_frame_id = source_frame;
 
   return msg;
@@ -85,6 +87,59 @@ BufferCoreROSConversions::lookupTransform(
     fixed_frame);
 }
 
+geometry_msgs::msg::VelocityStamped BufferCoreROSConversions::lookupVelocity(
+  const std::string & tracking_frame, const std::string & observation_frame,
+  const tf2::TimePoint & time, const tf2::Duration & averaging_interval) const
+{
+  // ref point is origin of tracking_frame, ref_frame = obs_frame
+  return lookupVelocity(
+    tracking_frame, observation_frame, observation_frame, tf2::Vector3(
+      0, 0,
+      0), tracking_frame, time,
+    averaging_interval);
+}
+
+geometry_msgs::msg::VelocityStamped BufferCoreROSConversions::lookupVelocity(
+  const std::string & tracking_frame, const std::string & observation_frame,
+  const std::string & reference_frame, const tf2::Vector3 & reference_point,
+  const std::string & reference_point_frame,
+  const tf2::TimePoint & time, const tf2::Duration & averaging_interval) const
+{
+  const tf2::Stamped<std::pair<tf2::Vector3, tf2::Vector3>> stamped_velocity = lookupVelocityTf2(
+    tracking_frame, observation_frame, reference_frame,
+    reference_point, reference_point_frame, time, averaging_interval);
+
+  geometry_msgs::msg::VelocityStamped msg;
+  const std::chrono::nanoseconds ns = std::chrono::duration_cast<std::chrono::nanoseconds>(
+    stamped_velocity.stamp_.time_since_epoch());
+  const std::chrono::seconds s = std::chrono::duration_cast<std::chrono::seconds>(
+    stamped_velocity.stamp_.time_since_epoch());
+  msg.header.stamp.sec = static_cast<int32_t>(s.count());
+  msg.header.stamp.nanosec = static_cast<uint32_t>(ns.count() % 1000000000ull);
+  msg.header.frame_id = reference_frame;
+  msg.body_frame_id = tracking_frame;
+
+  msg.velocity.linear.x = stamped_velocity.first.x();
+  msg.velocity.linear.y = stamped_velocity.first.y();
+  msg.velocity.linear.z = stamped_velocity.first.z();
+  msg.velocity.angular.x = stamped_velocity.second.x();
+  msg.velocity.angular.y = stamped_velocity.second.y();
+  msg.velocity.angular.z = stamped_velocity.second.z();
+
+  return msg;
+}
+
+bool BufferCoreROSConversions::setTransform(
+  const geometry_msgs::msg::TransformStamped & transform,
+  const std::string & authority, bool is_static)
+{
+  tf2::Stamped<tf2::Transform> tf2_transform;
+  tf2::fromMsg(transform, tf2_transform);
+
+  return setTransformTf2(
+    tf2_transform, transform.header.frame_id, transform.child_frame_id,
+    tf2_transform.stamp_, authority, is_static);
+}
 
 geometry_msgs::msg::TransformStamped
 Buffer::lookupTransform(
